# Speed up of classical UPS implementation