… slowquant/unitary_coupled_cluster/operators.py