The speed up is achieved through three different changes.

- An implementation of bitcount for Numba, which lets the performance-critical part of the code be JIT compiled.
- Changing the date structure of annihilation/creation operators to being tuples of integers and bools instead of class objects. This also allows for more direct indexing of det2idx.
- Implementation of in-place algebra for the FermionicOperator class. This avoids a lot of init calls and memory copies. 

Benchmark for wave function optimization using SLSQP with fUCCSD for Hydrogen chains with a distance of 0.8 Å.

| Active-space  | # parameters |  Old time per iteration (s) | New time per iteration  (s)|
|---|---|---|---|
| (2, 2)  | 3 |  3.1e-3 | 2.9e-3|
| (4, 4)  | 26 | 4.4e-2 | 3.3e-2 |
| (6, 6)  | 117 | 4.7e-1| 1.9e-1 |
| (8, 8)  | 360 | 10| 1.1 |
| (10, 10)  | 875 |  270 | 17 |
| (12, 12)  | 1818 | 7000 |  510 |
| (14, 14)  | 3381| >145000  | 17500  |